+++ cligen_syntax.h
@@ -84,10 +84,10 @@ int cligen_parse_line(cligen_handle h,
 	       cg_var       *arg,
 	       int           hide);
 
-int cligen_callback_str2fn(parse_tree, cg_str2fn_t *str2fn, void *arg);
-int cligen_callbackv_str2fn(parse_tree pt, cgv_str2fn_t *str2fn, void *arg);
-int cligen_expandv_str2fn(parse_tree pt, expandv_str2fn_t *str2fn, void *arg);
-int cligen_translate_str2fn(parse_tree pt, translate_str2fn_t *str2fn, void *arg);
+int cligen_callback_str2fn(parse_tree *pt, cg_str2fn_t *str2fn, void *arg);
+int cligen_callbackv_str2fn(parse_tree *pt, cgv_str2fn_t *str2fn, void *arg);
+int cligen_expandv_str2fn(parse_tree *pt, expandv_str2fn_t *str2fn, void *arg);
+int cligen_translate_str2fn(parse_tree *pt, translate_str2fn_t *str2fn, void *arg);
 int cligen_parse_debug(int d); 
 
 #endif /* _CLIGEN_SYNTAX_H_ */